HUD Visibility settings
-Lvl tree for Powerups and Waves in InfoGameMode(no drop chances yet)
-Default movement for Android is mouse with dtapShoot, new Shadowdash button for these settings
-Fixed Enemy Ship moving even after pausing

+++ SSS222/Assets/Scripts/SpanwersEvents/LootTableWaves.cs
@@ -8,7 +8,7 @@ public class LootTableEntryWaves{
     [HideInInspector]public string name;
     public WaveConfig lootItem;
     public float dropChance=0f;
-    public float levelReq=0f;
+    public int levelReq=0;
 }
 [System.Serializable]
 public class ItemPercentageWaves{
